fix : 자잘한 최적화 수정
